Add: advanced tasks
16. Everyone is a proponent of strong encryption:
Write a script that encodes and decodes text using
the rot13 encryption. Assume ASCII.

17. The eggs of the brood need to be an odd number:
Write a script that prints every other line from the
input, starting with the first line.

18. I'm an instant star. Just add water and stir:
Write a shell script that adds the two numbers stored in
the environment variables WATER and STIR and prints the result.
    - WATER is in base water
    - STIR is in base stir.
    - The result should be in base behlnort