Add: 0. How many subs?
Write a function that queries the Reddit API and returns the number
of subscribers (not active users, total subscribers) for a given
subreddit. If an invalid subreddit is given, the function should
return 0.

Hint: No authentication is necessary for most features of the Reddit
API. If you’re getting errors related to Too Many Requests, ensure
you’re setting a custom User-Agent.